Add winning panel, refactor
